Fixed types

+++ core/src/main/java/core/model/Location.java
@@ -21,7 +21,7 @@ public class Location {
     private Long id;
 
     @NotNull(message = "location.x not provided")
-    private Double x;
+    private long x;
 
     @NotNull(message = "location.y not provided")
     private Integer y;
